modify custom as 1.0.2.11

+++ etc/platform/bin/curlhttpgps.sh
@@ -6,17 +6,10 @@ x=""
 
 GPS_LOG_PATH="/opt/log/gps/"
 GPS_LOG="/tmp/gps_out.log"
-TEMP_GPS_LOG_PATH="/opt/log/gps/temp_gps.log"
-
+#TEMP_GPS_LOG_PATH="/opt/log/gps/temp_gps.log"
 GPS_EN="/etc/platform/conf/gps_en"
 
-MAC=$(cat ${FILE_REGISTER} | jq -j ".mac" | tr -t ":" "-")
-
-echo $MAC
-
-if [ -f ${GPS_LOG} ];then
-        rm ${GPS_LOG}
-fi
+MAC=$(cat ${FILE_REGISTER} | jq -j ".mac" | tr  ":" "-")
 
 i=0
 for file in $(ls ${GPS_LOG_PATH} |grep gps-)
@@ -27,42 +20,27 @@ done
 
 for upload in ${file_name[*]}
 do
-	cat ${GPS_LOG_PATH}${upload} >> ${TEMP_GPS_LOG_PATH}
-	rm  ${GPS_LOG_PATH}${upload}
-done
-
-if [ -f $TEMP_GPS_LOG_PATH ];then
-	x=$(cat ${GPS_EN})
-	gps_data=$(cat $TEMP_GPS_LOG_PATH)
-	status=`curl --max-time 180  -F "type=gps" -F "signature=${x}" -F "ident=${MAC}" -F "content=${gps_data}"  -o  ${GPS_LOG} -s -w  %{http_code}   http://update1.9797168.com:821/wifibox/`
-#	while true
-#	do
-#       	if [ ! -f ${GPS_LOG} ];then
-#                	sleep 2
-#        	else
-#                	break
-#        	fi
-#	done
-	if [ $status -eq "200" ];then
-		outcontent=$(cat ${GPS_LOG} | jq -j ".success")
-        	echo out=$outcontent
-        	case "$outcontent" in
-                	false)
-                        	echo "GPS log upload error !---time is:"`date` >>/tmp/error.log
-                        	;;
-                	true)
-                        	rm $TEMP_GPS_LOG_PATH
-				echo "GPS log upload success !---time is:"`date`>>/tmp/error.log
-                        	;;
-                	*)
-                        	echo "Unknown status !---time is:"`date`>>/tmp/error.log
-                        	;;
-        	esac
-
-	else
-		 echo "Network have some unknown problems !---time is:"`date`>>/tmp/error.log
+	if [ -f ${GPS_LOG} ];then
+        	rm ${GPS_LOG}
 	fi
-else
-        echo "file do not exit!--time is:"`date` >>/tmp/error.log 
-fi
+        x=$(cat ${GPS_EN})
+        gps_data=$(cat ${GPS_LOG_PATH}${upload})
+        status=`curl --max-time 180  -F "type=gps" -F "signature=${x}" -F "ident=${MAC}" -F "content=${gps_data}"  -o  ${GPS_LOG} -s -w  %{http_code}   http://update1.9797168.com:821/wifibox/`
+        if [ $status -eq "200" ];then
+                outcontent=$(cat ${GPS_LOG} | jq -j ".success")
+		if [ ${outcontent} == "false" ];then
+			echo "${upload} GPS log upload error !---time is:"`date` >>/tmp/error.log
+			break
+		elif [ ${outcontent} == "true" ];then
+			rm  ${GPS_LOG_PATH}${upload}
+                        echo "${upload} GPS log upload success !---time is:"`date`>>/tmp/error.log
+		else
+			echo "${upload} Unknown status !---time is:"`date`>>/tmp/error.log
+			break
+		fi
+        else
+                 echo "Network have some unknown problems !---time is:"`date`>>/tmp/error.log
+        	break
+	fi
+done
